aggiunto controllo di sessione per bloccare l'accesso all'api se non sei connesso.

+++ server/routes/profile.js
@@ -7,6 +7,9 @@ const bodyParser = require('body-parser');
 module.exports = function (logger, app, io) {
 
     app.post('/user/:id', async (req, res, next) => {
+        if (req.session.user == null){
+            return res.send({success: false});
+        }
         let fullpath = './tmp/' + req.params.id + '.png';
         let bitmap = new Buffer(req.body.avatar, 'base64');
         fs.unlink('path/file.txt', (err) => {});
